refactor: :recycle: Modified the component to match our standards
I removed the default prop as well as prop validation using prop-types since we are already using TS types to validate props. I also moved the documentation of each prop to the type declaration to make it more consistent. Also changed the title and name of the component to be in sync with the Atomic Design principles that the comp lib uses for structure.