feat(community): add display of topics in thread component
The thread component now displays the topics associated with a thread. If a thread has topics, they will be displayed in a row below the thread title. The topics are displayed as a series of tags, with each tag representing a topic. This change improves the user experience by providing more context about the thread.