feat(community): updated home page to show author in thread component
The threads query now includes the topics field, which allows for displaying the topics of a thread. The showAuthor prop in the thread component is now set to true, which displays the author of the thread. This improves the user experience by providing more information about the thread and its author.