feat(community): Added topics to main threads (#119)
* feat(community): add display of topics in thread component

The thread component now displays the topics associated with a thread. If a thread has topics, they will be displayed in a row below the thread title. The topics are displayed as a series of tags, with each tag representing a topic. This change improves the user experience by providing more context about the thread.

* Automated formatting changes

* feat(community): updated home page to show author in thread component

The threads query now includes the topics field, which allows for displaying the topics of a thread. The showAuthor prop in the thread component is now set to true, which displays the author of the thread. This improves the user experience by providing more information about the thread and its author.

* feat(types): add topics field to ThreadType
The ThreadType now includes a topics field, which is an array of strings. This allows threads to be categorized by topic, which can be useful for filtering and searching.

---------

Co-authored-by: Dániel B. Papp <chef-danny-d@users.noreply.github.com>