refactor(thread): Optimized data layer of component using prop drilling (#120)
* refactor(community): add type-safety to comment hierarchy generator

The commentGen function now has a ThreadType parameter for parentComment and returns an array of ReactNode components or an empty array. This improves the code's readability and maintainability by making it easier to understand the function's purpose and expected inputs/outputs.

* refactor(thread): optimized data layer

The upvotes prop was not being used in the component, so it was removed along with the related code. The upvotesProp prop was added to display the number of upvotes the thread has in the upvote button.

* feat(thread): add upvotes to thread and comments
This commit adds the upvotes property to the thread and comments in the thread page. The upvotes property is an array of objects containing the id of the user who upvoted the thread or comment. The upvotesProp property is passed to the Thread component in the communities index page to display the number of upvotes for each thread.

* Automated formatting changes

---------

Co-authored-by: Dániel B. Papp <chef-danny-d@users.noreply.github.com>