feat(communities): add sorting and filtering functionality to threads home page (#122)
* feat(communities): add sorting and filtering functionality to threads list
This commit adds the ability to sort the list of threads by newest, oldest, most watched, and most active in the past 7 days. It also adds the ability to filter the list of threads by topic. These features improve the user experience by allowing them to quickly find the threads they are interested in.

* feat(input): add icon prop to component

The icon prop is added to the Input component to allow for the rendering of a search icon on the input element. This improves the user experience by providing a visual cue for the search functionality.

* chore(tsconfig): update target to es2016

The target version of the TypeScript compiler has been updated to es2016. This allows the use of newer language features and improves compatibility with modern browsers.

* Automated formatting changes

---------

Co-authored-by: Dániel B. Papp <chef-danny-d@users.noreply.github.com>