feat(type): add LessonProgress
The LessonProgress type is added to represent the progress of a user in a lesson. It contains the id of the lesson, the status of the lesson, whether it is completed or not, and the date it was created and updated. The LessonByModuleEnrollment type is added to represent a lesson with its progress. It contains the id, name, and position of the lesson, and an array of LessonProgress.