feat(module): add logic to determine if module has been started
The unused user data fetching has been removed to simplify the code. The module page now determines if the module has been started by checking if any of the lessons in the module have a lesson progress status that is not equal to 0. If any lesson has a non-zero status, the module is considered started. This logic is used to determine whether to display the "START MODULE" or "RESUME MODULE" button.