refactor(user): Removed duplicate & obsolete profile component (#147)
* refactor(user): Removed duplicate & obsolete profile component

* chore(user): Moved profile component & story

* chore: Automatic formatting changes

* feat(type): Updated user to include social object

* chore: Created aliased path for types directory

* chore: Update project version and name

* refactor(user): Started extraction of business logic

* Automated formatting changes

* feat(user): Converted page to use profile component

* Automated formatting changes

* feat(request_failed): add RequestFailed component and its story
The RequestFailed component is a new component that displays an error message when a request fails. It has two props, title and subtitle, which are used to customize the error message. The component also includes a button that allows the user to try the request again. The RequestFailed component is now available for use in the application.

* refactor(user_profile): replace error message with RequestFailed component
The error message displayed when user information could not be retrieved has been replaced with the RequestFailed component. This component provides a more consistent and visually appealing way of displaying error messages.

* Automated formatting changes

---------

Co-authored-by: Dániel B. Papp <chef-danny-d@users.noreply.github.com>