feat(plan): add plan of study index page (#149)
* feat(sidebar.tsx): add user-specific plan of study link
The href attribute of the "PLAN OF STUDY" sidebar item is now dynamically generated based on the user's openId. This allows for a more personalized experience for the user as they can easily access their own plan of study.

* feat(plan): add plan of study index page
This commit adds a new page for the plan of study feature. The page displays four panels, each with a title and a body. The panels are generated using a map function and a new component called PlanOfStudyPanel. The page also uses the Layout component from the common pages layouts folder. The page is only accessible to authenticated users, and it greets the user by name.

---------

Co-authored-by: Dániel B. Papp <chef-danny-d@users.noreply.github.com>